+++ src/Templates/Npm.fs
@@ -1,0 +1,7 @@
+namespace RunHelpers.Templates
+
+open RunHelpers.BasicShortcuts
+
+[<RequireQualifiedAccess>]
+module Npm =
+    let install () = npm [ "install" ]