+++ src/Templates/Pnpm.fs
@@ -1,0 +1,7 @@
+namespace RunHelpers.Templates
+
+open RunHelpers.BasicShortcuts
+
+[<RequireQualifiedAccess>]
+module Pnpm =
+    let install () = pnpm [ "install" ]